Completado Test para el controlador (mejorable)

+++ perreraWS/src/main/java/com/ipartek/formacion/perrera/controller/PerroController.java
@@ -55,7 +55,7 @@ public class PerroController {
 	@ApiResponses(value = { @ApiResponse(code = 200, message = "Todo OK"),
 			@ApiResponse(code = 204, message = "No existe perro con esa ID"),
 			@ApiResponse(code = 500, message = "Error inexperado en el servidor") })
-	public Response getById(@PathParam("id") int idPerro) {
+	public Response getById(@PathParam("id") long idPerro) {
 
 		try {
 			PerroDAOImpl dao = PerroDAOImpl.getInstance();
@@ -78,7 +78,7 @@ public class PerroController {
 	@ApiResponses(value = { @ApiResponse(code = 200, message = "Perro eliminado"),
 			@ApiResponse(code = 204, message = "No existe Perro con ese ID"),
 			@ApiResponse(code = 500, message = "Error inexperado en el servidor") })
-	public Response delete(@PathParam("id") int idPerro) {
+	public Response delete(@PathParam("id") long idPerro) {
 
 		try {
 			PerroDAOImpl dao = PerroDAOImpl.getInstance();
@@ -127,7 +127,7 @@ public class PerroController {
 			@ApiResponse(code = 204, message = "No existe perro con ese ID"),
 			@ApiResponse(code = 409, message = "Perro existente, no se puede modificar"),
 			@ApiResponse(code = 500, message = "Error inexperado en el servidor") })
-	public Response put(@PathParam("id") int idPerro, @PathParam("nombre") String nombrePerro,
+	public Response put(@PathParam("id") long idPerro, @PathParam("nombre") String nombrePerro,
 			@PathParam("raza") String razaPerro) {
 		try {
 			PerroDAOImpl dao = PerroDAOImpl.getInstance();
